implementation of Android-specific Ti.Filesystem.applicationCacheDirectory TIMOB-8685

+++ android/titanium/src/java/org/appcelerator/titanium/io/TiFile.java
@@ -93,7 +93,10 @@ public class TiFile extends TiBaseFile
 		boolean deleted = true;
 
 		File[] files = d.listFiles();
-		
+		if (files == null) {
+			return false;
+		}
+
 		for (File f : files) {
 			if (f.isFile()) {
 				deleted = f.delete();
